File structure change: Moved controls to their own folder to keep tidy
