+++ BuldingMapper/BuildingMapper.csproj.user
@@ -1,19 +1,19 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
-    <Compile Update="FloorEditor.cs">
+    <Compile Update="Controls\FloorEditor.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Update="MainForm.cs">
+    <Compile Update="Forms\MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="NewBuildingForm.cs">
+    <Compile Update="Forms\NewBuildingForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="RoomEditorForm.cs">
+    <Compile Update="Forms\RoomEditorForm.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Update="RoomEditorPanel.cs">
+    <Compile Update="Controls\RoomEditorPanel.cs">
       <SubType>Component</SubType>
     </Compile>
   </ItemGroup>
