+++ BuldingMapper/Controls/FloorEditor.resx
