+++ BuldingMapper/Controls/RoomEditorPanel.resx
